Refactor to new structure, start H7 modifications
