+++ Cargo.toml
@@ -5,5 +5,6 @@ members = [
 ]
 exclude = [
   "boards/stm32l476rg",
+  "boards/stm32h743zi"
 ]
 resolver = "2"
