+++ boards/stm32h743zi/src/lib.rs
@@ -1,0 +1,4 @@
+#![no_std]
+
+pub mod log;
+pub mod tasks;
