+++ boards/stm32h743zi/src/tasks.rs
@@ -1,0 +1,2 @@
+pub mod mqtt_send_and_recv;
+pub mod test_mqtt_tasks;
